refactors network lib to make it version independent
