+++ service/controller/v4/cloudconfig/cloud_config.go
@@ -17,7 +17,7 @@ import (
 	"github.com/giantswarm/azure-operator/client"
 	"github.com/giantswarm/azure-operator/service/controller/setting"
 	"github.com/giantswarm/azure-operator/service/controller/v4/key"
-	"github.com/giantswarm/azure-operator/service/controller/v4/network"
+	"github.com/giantswarm/azure-operator/service/network"
 )
 
 type Config struct {
