+++ service/controller/v4/resource_set.go
@@ -3,6 +3,7 @@ package v4
 import (
 	"context"
 	"fmt"
+	"net"
 	"time"
 
 	"github.com/giantswarm/apiextensions/pkg/clientset/versioned"
@@ -24,7 +25,6 @@ import (
 	"github.com/giantswarm/azure-operator/service/controller/v4/credential"
 	"github.com/giantswarm/azure-operator/service/controller/v4/debugger"
 	"github.com/giantswarm/azure-operator/service/controller/v4/key"
-	"github.com/giantswarm/azure-operator/service/controller/v4/network"
 	"github.com/giantswarm/azure-operator/service/controller/v4/resource/deployment"
 	"github.com/giantswarm/azure-operator/service/controller/v4/resource/dnsrecord"
 	"github.com/giantswarm/azure-operator/service/controller/v4/resource/endpoints"
@@ -35,6 +35,7 @@ import (
 	"github.com/giantswarm/azure-operator/service/controller/v4/resource/service"
 	"github.com/giantswarm/azure-operator/service/controller/v4/resource/vnetpeeringcleaner"
 	"github.com/giantswarm/azure-operator/service/controller/v4/resource/vpngateway"
+	"github.com/giantswarm/azure-operator/service/network"
 )
 
 type ResourceSetConfig struct {
@@ -362,7 +363,11 @@ func NewResourceSet(config ResourceSetConfig) (*controller.ResourceSet, error) {
 			return nil, microerror.Mask(err)
 		}
 
-		subnets, err := network.ComputeFromCR(ctx, azureConfig)
+		_, vnet, err := net.ParseCIDR(key.VnetCIDR(azureConfig))
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+		subnets, err := network.Compute(*vnet)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
